+++ checkpoint/orbax/checkpoint/path/deleter_test.py
@@ -40,7 +40,7 @@ class CheckpointDeleterTest(parameterized.TestCase):
         primary_host=None,
         directory=self.ckpt_dir,
         todelete_subdir=todelete_subdir,
-        name_format=step_lib.StandardNameFormat(),
+        name_format=step_lib.standard_name_format(),
         enable_background_delete=False,
     )